src/requirements.txt
Django==2.1.1
django-filter==2.0.0
django-rest-framework==0.1.0
djangorestframework==3.8.2
pytz==2018.5
